Use ULTIMATE_ID instead of ALIAS_ID
When using this Query for Cedant 09850C1 01 to compare 2020Q1 and 2020Q4, I noticed that there were some ALIAS_ID that changed, but the parent stayed the same. 
That resulted in the query result not bringing in all the exposures' ECap.
Therefore, I'll change the query to use ULTIMATE_ID as to avoid this scenario in the future.